Update README.md
WebGPU is a cutting-edge web standard that enables developers to leverage the full capabilities of modern GPUs for graphics and compute tasks within web applications. To streamline the integration of <a href="https://gpuweb.github.io/gpuweb/explainer/">WebGPU</a> into web development as a components library to the client GPU it is useful to call an abstraction layer called ".ddd." This proposal presents an overview of the .ddd abstraction, explaining its role as a mediator between WebGPU components and the application. By encapsulating complex WebGPU functionalities, .ddd simplifies development, enhances maintainability, and ensures cross-platform consistency.